refactor: Move file tree to components
