+++ frontend/src/components/file-tree/__tests__/utils.test.ts
