+++ frontend/src/components/file-tree/types.ts
