+++ frontend/src/components/file-tree/utils.ts
